make view more consistent with index types

+++ src/tensor/tensor_ops.rs
@@ -676,7 +676,7 @@ impl<T: Copy> Tensor<T> {
     pub fn var(&self) -> f32 {
         unimplemented!()
     }
-    pub fn view(&self, dims: &[i32]) -> Self {
+    pub fn view(&self, dims: &[usize]) -> Self {
         unimplemented!()
     }
     pub fn view_as(&self, tensor: &Self) -> Self {
@@ -1370,7 +1370,7 @@ impl TensorKind {
     pub fn var(&self) -> f32 {
         unimplemented!()
     }
-    pub fn view(&self, dims: &[i32]) -> Self {
+    pub fn view(&self, dims: &[usize]) -> Self {
         unimplemented!()
     }
     pub fn view_as(&self, tensor: &Self) -> Self {
